Removes unused Failure classes inheriting from Belos::Error
There are many Failure classes that are defined in the iteration
classes and solver managers that are not used.  Many of these are
the result of general code copying during the development of new
solvers.  However, if they are not being used, they should be
removed to improve the general clarity of the code.