Ifpack2: fix build+test with Ifpack2_ENABLE_DEBUG
Fix the CMake logic for this option, and the option for the QD TPL.
The cache variables that control cmakedefines in config headers
put the package name in all caps ($PACKAGE_NAME_UC in tribits).

Fix a bug in LocalFilter debug checks - the map isLocallyFitted checks
were backwards. If map2.isLocallyFitted(map1), map1 is a subset of map2,
not the other way around. For example, colMap.isLocallyFitted(domainMap) is
typically true but domainMap.isLocallyFitted(colMap) is not.
Added a comment to this function in Tpetra::Map to make this easier to understand in future.

Fix a build error in a debug block of TriDiContainer.