Switch over to using FindTPL<tplName>Dependencies.cmake files (TriBITSPub/TriBITS#494)
Now no errors will occur when all of the upstream TPLs are not enabled.  For
exmaple, this will allow ParMETIS to be enabled but not METIS (like is done
with the Trilinos PR GenConfig builds).

I also removed item from RELEASE_NOTES on break in backward compatibility for
TPL dependencies.