Check and abort if existing subdirs are not git repos (#10766)
This checks to make sure that an existing srn-ini-files or son-ini-files
subdir is a Git is a git repo or not and abort right away if it is not.

I also added -e to the bash invocation to make sure it aborts on any error.

This will catch the case where an existing subdir is left over from the old
submodules implementation where someone ran it locally and it generated
symlinks.  This left over a non-git subdir and broke the behavior of the
script (see #10766).