Disable pseudo package ShyLU in ATS-2 builds (#10865)
This should avoid a strange find_package(Trilinos) error being generated about
missing ShyLUConfig.cmake.  But since ShyLU is not a real TriBITS package
(i.e. does not contain any libs, tests, or subpkages and there are no
downstream dependencies on the pseudo package ShyLU), this will cause no lack
of testing.

The right solution will be to remove the pesudo ShyLU package and instead
implement it as a TriBITS meta-package (see #9497).