Disable Pliris tests in ATS2 GenConfig builds (#10931)
For some reason, CMake is not respecting the force disable of
Trilinos_ENABLE_Pliris from the last commit.  Rather than try to debug this, I
am just disabling the tests which will allow PR #10930 to pass the PR builds
and merge.

There are other issues with the GenConfig files that will need to be addressed
that I am seeing.