Galeri: copy missing header during installation
This change includes the file galeri/src-eptra/Galeri_Exception.h into
the list of headers of Galeri's xpetra portion to be copied to the
include directory during installation.

This is only a quick fix to hopefully get Galeri working for the
upcoming EuroTUG 2022 meeting. A long-term solution coudl be to move the
header outside of `src-epetra` to a place, that's independent of Epetra
or Xpetra.

Part of #10996.