Ifpack2: fix performance bug in OverlappinRowMatrix
OverlappingRowMatrix needs to keep track of unique GIDs to add to
the column map of the next overlap level. It was doing this with linear
search in an unsorted array, for O(N^2) time. Instead, use unordered_set
to keep track of unique GIDs to reduce this part to O(N).