Tpetra TSQR: fix lwork in LAPACK tester
compute_QR and compute_explicit_Q each have their own function to
compute the work length (lwork). Since the TSQR test uses the same work
array for both computations, its length needs to be the max of the two lworks.

There is already a function lworkQueryLapackQr to do this max, but it
can't be used with cuSOLVER because cusolverDnXorgqr_bufferSize needs
the actual A/tau pointers.

Fixes #10847.