Sacado:  Fix memory leaks in Rad when using DFad nested scalar type.
Rad does its own memory management, but doesn't know anything about it
when its nested scalar type is also dynamically allocated, which was
causing massive memory leaks when using, e.g., Rad< DFad<double> >.  The
only way to fix this without a major overhaul in Rad is to ask the user
to call a new function to zero out all of Rad's internally stored data,
which as a side-effect will deallocated DFad derivaive arrays.  Also
added this new function to the trad-dfad example.

For issue #7741.