Ifpack2: Update PowerMethod to compute absolute eigenvalue or eigenvalue
The Dinv*A matrix for this problem has two dominant eigenvalues, 12.5839 and -10.6639. The original implementation used a random guess and computed the eigenvalues in iterations using the norm, but returned the final answer using xAx. This caused spurious issues where the history would appear to converge, but the final answer would be horribly incorrect. Now the power method computes the spectral radius by default, and it computes xAx iteratively if compute spectral radius is set to false