Piro: refactor to better support adjoint solves
Adjoint solves require the application of the inverse of the jacobian adjoint.
Not all the solvers support transpose solves, so we allow the user to pass an adjoint model evaluator
where the user can provide a customized solve (e.g. by assemblying an adjoint jacobian).
This mimics whats done for transient problems.