Adds HAVE_TEUCHOS_COMPLEX ifdef to specializedLAPACK<std::complex<T> >::test
The specializedLAPACK<std::complex<T> > class is conditionally compiled
based on whether complex-valued arithmetic is enabled.  The implementation
of the specializedLAPACK<std::complex<T> >::test method was not placed
behind an #ifdef HAVE_TEUCHOS_COMPLEX when it should have been.  This was
causing compilation errors on machines where complex-valued arithmetic
is not enabled.

https://github.com/trilinos/Trilinos/issues/11168