Kokkos-Kernels:  Disable compile-time on support scalar types for cuSPARSE
This change removes the compile-time static-assert on valid scalar types
for the cuSPARSE TPL.  The run-time check is still there, and this
mimics the behavior for the interface to the older (CUDA-10) cuSPARSE
interface.  Making this a compiler error makes it very difficult for
packages such as Ifpack2 to support scalar types that cuSPARSE might
not, since it can't even instantiate Kokkos-Kernels when cuSPARSE is
eanbled and fixes several compiler errors when Ifpack2 is instantiated
on Stokhos scalar types.