Remove build_name changes from LaunchDriver.py
LaunchDriver.py is untested in the current PR framework, since it's
using the target branch at that point instead of the source branch (or
the merged branch).  Want to eliminate as much stuff as possible from
this script.  Most stuff doesn't need to be set until later anyway.

User Support Ticket(s) or Story Referenced: TRILFRAME-528