Change build core calculation
Calculate and use the build cores inside of PullRequestLinuxDriver.sh,
so that the PR testing will make use of this setting and test it.  In
response, also stop using TRILINOS_MAX_CORES to make it more clear that
nothing above this script can change the build core count.

User Support Ticket(s) or Story Referenced: TRILFRAME-528