Revert "Only use 1 slot per GPU"
This reverts commit 1f03b84ab6e3eed40be413ff733d9d88ad8e9116.

Sometimes the GPU machines have a GPU go offline (talking with the
admins about that), and that means if we only use 1 slot per GPU, we'll
run out of total slots to run the 4-rank MPI tests.  It turns out the
errors with Kokkos were NOT due to the 2-slot assignment, so let's go
ahead and use 2 slots.