SEACAS: Remove uses of fmt in include files
SEACAS uses the lib{fmt} library for output, but that library is
not yet a Trilinos TPL, so external packages using Trilinos
are not required to have the lib{fmt} include files available.

SEACAS in Trilinos uses lib{fmt} as a header-only library and there
is an embedded version in SEACAS/Trilinos.

This commit removes all fmt-related includes from header files so
applications using Trilinos do not need to have the fmt include files.