TrilinosCouplings: resolve shadow warning
This actually appears to be a bug.  It appears that the variable
ratio was intended to be modified within a for-loop, and then used
afterwards.  However, it was declared within the loop (hence the shadow),
and thus never changing the outer declared variable's value.

I also replaced a bunch of tabs with spaces.