Move check for ParMETS version for Zoltan2 to Zoltan2 (#63)
When ParMEITS is found by an upstream cmake project (e.g. KokkosKerenls), the
value of HAVE_PARMETIS_VERSION_4_0_3l does not get written into the
ParMETISConfig.cmake file so Zoltan2 was configuring with an error.  But this
is a check for Zoltan2, not other Trilinos packages anyway so this check
should be in Zoltan2, not in the FindTPLParMETIS.cmake

NOTE: If we want to support exporting various variables into the generated
<tplName>Config.cmake files for TriBITS TPLs generated using
tribits_tpl_find_include_dirs_and_libraries(), then we will need to do some
refactoring in TriBITS to make that possible. But that should not be too hard.
This was just not needed in this case.