TrilinosInstallTests_find_package_Trilinos: Run in own subdir
There was a mistake in the defintion of the test
TrilinosInstallTests_find_package_Trilinos in that it created a dummy CMake
project right from within the package directory:

  packages/TrilinosInstallTests/

When I changed the name of a base dir locally, the test failed showing the
error:

CMake Error: The current CMakeCache.txt directory <common-base>/demo-external-kokkos/Build_Trilinos/packages/TrilinosInstallTests/CMakeCache.txt is different than the directory <common-base>/Build_Trilinos/packages/TrilinosInstallTests where CMakeCache.txt was created. This may result in binaries being created in the wrong place. If you are not sure, reedit the CMakeCache.txt

Now the dummy project dir is in a new subdir that is deleted and created anew
each time the test runs.