Panzer: release local views to avoid tpetra throw
Thought we had all of these out of the code, but running at scale on
cuda has revealed some new ones. Tpetra switches execution spaces
based on problem size for some operations (and whether cuda aware mpi
is enabled), so this only occurs if on cuda with a big enough
problem. Not sure how to test this without adding special flags to
tpetra to exercise the path for small problems.