Add hook for inserted package TrilinosLinearSolvers (#11730)
We can't add the package TrilinosLinearSolvers until it goes through
copyright.

NOTE: This is being added as an 'PT' package even through it does not exist in
a fresh clone.  That is fine since this will not impact Trilinos PR testing
until the package gets added to the repo proper.