Increase parallel link limit to 8
Now that we're using ccache, we shouldn't be running large numbers of
compiles, so we should have more resources for linking.  Also, we're
more bound by the link time now, so use more parallelism.  If it's a
"good" cache build with lots of hits I've observed the machine load to
be very low now, so hopefully we can take advantage of this to enhance
the speedup.

User Support Ticket(s) or Story Referenced: TRILFRAME-521