Stratimikos: Add offical support for Ifpack2 for Tpetra matrices (#11730)
NOTE: There was a lot of existing code in various packages that was
registering an Ifpack2 preconditioner factory.  I just remove the registration
to use the Ifpack2 preconditioner factory already registered.  The removed a
lot of redundant code.

We need to do the same for MueLu as well (after switching to static/global
registration of new linear solvers and preconditioner factories).

These changes were done as part of #11730