Stratimikos: BelosLinearOpWithSolve: Have VERB_LOW provide description() (#11730)
Before, passing in a VERB_LOW did not print anything.  Now it provides the
description() value.  (But there is likely too much detail in the
description() string since this is meant to be a shorter one-line name.)

Also, VERB_HIGH and higher will give the longer multi-line describe() output.
Before, it was only provided by VERB_EXTREME.  But VERB_EXTREME is meant to
provide all matrix and vector values and extreme detail.  Passing in a verb
level below VER_EXTREME will not print all that info since the verb level is
passed down to the aggergate objects.

These changes were done as part of #11730.