MV::means: The implementation needs an additional fence
Currently the values coming out of KokkosBlas::sum() are
assumed to be avaiable after the function returns for MPI
communication. However that Kokkos Kernels functions is
non blocking hence might still be runnning on the device
while the host function call returns. We need to fence to
guarantee that the device function has completed. Before
performing the MPI reduction.