Xpetra: Change to call getConstTpetraOperator() to address const issue (#11730, #11731)
When you give the Thyra::MueLuPreconditionerFactory a proper const
Tpetra::Operator object, there is code in Xpetra that incorrectly trys to grab
a non-const RCP, which showns an exception.  But there is code that calls
rcp_const_cast() anyway to cast away const anyway so what is the point of
asserting non-const?  (See #11731.)

The workaround in this commit is to change from calling getTpetraOperator() to
getConstTpetraOperator().

These changes were done as part of #11730 and see #11731.