Teuchos: Add throwNumber to TestForException_break() (#11730)
This allows you to set a condition in a debugger like GDB:

  cond <break-point-num> thrownNumber==<throwNumber>

This is what is documented in the Teuchos Memory Management guide is that you
can set this breakpoint.  That must have broken over 10 years ago.

I also updated the Doxygen documentation to note setting a condition to break
on throwNumber.

The ShyLU_DDFrosch macro FROSCH_ASSERT() had to be updated as well and could
not just call TEUCHOS_TEST_FOR_EXCEPTION() because of the different
indentation.

These changes were done as part of #11730, to make it easier to debug some
exceptions.