Explicitly split input string replacement list by whitespace (#11545)
I am not sure what happened but this Python script but it was no longer
breaking up the read in list of tokens into individual array entries by
whitespace.  Therefore, I add to add some code to make that split explicit.

I manaully tested this and it seemed to work okay.