Tpetra: fix #11820
In #11679, the WDV "sync path" was disabled when in a host-parallel region.
However, it turns out that Serial().in_parallel() always returns false, so
parallel_ regions running on Serial would suffer the overhead of WDV tracking.
This brings back the explicit enable/disable of WDV tracking.
This means a few more lines of boilerplate, but it fixes perf regression #11820,
and in fact makes fillComplete in FEAssembly 20% faster than before #11679.