Manually remove redundant Kokkos dep (#11545)
Somehow, the script remove_kokkos_subpackages_r.sh did not remove this
redundancy.  (This is likely because Kokkos was at the end of the line with no
space and no ')' char .)  But this looks to be the only case in Trilinos it
appears where redundant 'Kokkos Kokkos' was not removed by that script.