Minor cleanup to address build warnings and memory leaks
The new dashboard for memory leaks in Trilinos has identified
an issue in the Anasazi utiltity that generates test problems
for the eigensolver.

https://trilinos-cdash.sandia.gov/index.php?project=Trilinos#!/%23DynamicAnalysis

While checking this issue, I noticed a build warning in the solver
factory and potential issues in building the Thyra examples and tests.

This commit addresses all these minor issues.