KokkosKernels: Don't list include for non-existant 'batched' build dir (#11966)
See the embedded comment.  This fixes the generated KokkosKernelsTargets.cmake
file in the build dir when tests are not enabled.

NOTE: This is a very non-conventional CMake build system.  Your never see
CMakeLists.txt files that are included with include().  These should always be
pulled in using add_subdirectory().  If you want to just include() some CMake
code, you call it `<SomeLogicalBaseName>.cmake` and then include() it.