Intrepid2: Support Orientations with Structured Integration (#11667)
Intrepid2: Support Orientations with Structured Integration (#11667)

Structured integration to date (i.e., support for sum factorization in tensor-product bases, with the hope to extend to support for matrix-free computations) has not explicitly supported application of element orientations; assembly has been performed locally, with global assembly entirely left to the user. (Existing support for orientations in Intrepid2 has assumed application of orientations to basis values, effectively erasing the structural information required for sum factorized assembly.)

This PR addresses this by adding support for application of orientations to an assembled local matrix. The workflow required is as follows:

1. Record orientations into a (C) View of orientations, where C is the cell index. (See CellGeometry::orientations().)
2. Perform local assembly into a (C,F1,F2) View, where F1 is the field index for the left basis, and F2 is the field index for the right basis. (See IntegrationTools::integrate().)
3. Call the new OrientationTools::modifyMatrixByOrientation(), passing in the local, unoriented (C,F1,F2) matrix, the orientations, and the left and right basis. This will store into an oriented (C,F1,F2) matrix.

It's worth noting that in the local assembly, one should use the local/unoriented basis for everything. If doing a nonlinear problem, then the basis coefficients used there should be expressed not in terms of the global solution (oriented basis), but in terms of the local basis.

To transform global coefficients (on an oriented basis) into local coefficients (on an unoriented basis) is to apply the transpose of the orientation matrix that took you from local basis values to global basis values. This can be accomplished using the new OrientationTools::modifyBasisByOrientationTranspose().

This PR also adds the following:
- Made all the tests comparing standard to structured assembly in test group StructuredIntegration use orientations, so that the new orientation support is exercised in all those tests
- Added a more general, asymmetric mechanism for performing both standard and structured assembly of one basis function integrated against another, in assembly-examples/StandardAssembly.hpp and assembly-examples/StructuredAssembly.hpp, respectively.
- Added a host of asymmetric tests in the test group StructuredIntegration, against 1D, 2D, and 3D examples and against all function spaces, and a few polynomial order combinations (p=1 and p=2).
- Split StructuredIntegrationTests.cpp into seven files (five .cpp files, and two .hpp files) for cleaner organization and faster compilation.
- Resolved a couple bugs in the kernels for IntegrationTools::integrate() revealed by the new tests.
- In Data:
  - in storeInPlaceCombination(): fixed the ArgExtractor for a case when B (not B_underlying) is used; was incorrectly using the underlying ArgExtractor.
  - added a constructor with run-time logical rank (but compile-time View rank)
  - fixed the setUnderlyingViewN() methods to be non-const
  - fixed shallowCopy() to properly set the rank of its returned Data object
- Added DataTools:
  - Utility methods for manipulating Intrepid2::Data objects
  - Initial method is multiplyByCPWeights() which allocates, fills, and returns a Data container that is the result of a (C,P,D,D) container weighted by values from a (C,P) container.
- In TransformedBasisValues, added multiplyByPointwiseWeights() method to modify the transformed basis values by user-supplied (C,P) weights
- In CellTools, added method setJacobianDividedByDet()