Chomp usage whitespace before assertion
I absolutely do not care how the usage is formatted.  I barely care
what's in it, given that it's built into the Python argparse library.
But, the test is here, so here it shall stay for now.  Basically run a
chomp() on it prior to asserting equal to avoid lots of whitespace diff
headaches.

User Support Ticket(s) or Story Referenced: TRILFRAME-573