Working on H^1 Pyramid implementation: - added missing definitions of mu, nu functions and their gradients - added a basis equivalence test comparing the existing lowest-order H^1 pyramid implementation to the new one - tweaked CMakeLists.txt to honor the Pyramid ETI for double, Kokkos::DefaultExecutionSpace
With those changes, the code compiles.  However, the pyramid implementation is incomplete; e.g. we don't have tags set up correctly, and there are some value computations that are still copied and pasted from the tet implementation.  The basis equivalence test fails with a Kokkos View out-of-bounds error inside IntegratedLegendreBasis_HGRAD_PYR; it's a member view that is being accessed, and probably the code sizing this is part of the unrevised construction code.