Intrepid2: Add H^1 Hierarchical Pyramid Basis (#12079)
## Motivation
This PR adds an arbitrary-order, hierarchical basis for H(grad) on the pyramid.  This follows #11061, which added hierarchical bases for the wedge, and  #11009, which did so for the tetrahedron.  We will be adding a hierarchical L^2 basis for the pyramid soon after this, followed by H(div) and H(curl).

We also add support for generating equispaced point lattices on the pyramid.

## Testing
A basic "basis equivalence" test, which establishes that the new basis spans the same space as the existing low-order nodal basis, is included here.  Additionally, I have done offline comparison with the ESEAS implementation of this basis, testing up to 9th order, with excellent agreement.  I hope to include these tests with Intrepid2 soon; this has been prevented previously by the license for ESEAS, but they are changing the license to one that will allow inclusion.

In the follow-on PRs, we also plan to add tests against orientations and projection with the new bases, as well as tests against the association of basis degrees of freedom with subcell topologies.