Intrepid2: modify Largangian Interpolation implementation
We use inverse of orientation maps to simplify the implementation of Lagrangian Interpolation.
A major advantage is that now we provide an interpolation method that requires the
evaluation of the function on the reference dof nodes, which are the same in each cell,
instead of the "oriented" dof nodes that change from cell to cell.
This has the advantage of evaluating the basis functions only on the reference cell instead of on every cell.