Intrepid2: refactor of projections using new symmetric cubature rules
The use of symmetric cubature rules allow to require the evaluation of the
target function at points in the reference cell, which are the same in each cell,
instead of requesting the evaluation of the function at oriented quadrature points,
which can be different in different cells.

We refacor the projections so that the evaluation points are now stored in the ProjectionStruct
and can be accessed using the functions getAllEvalPoints() and getAllDerivEvalPoints.
And we change the implementation of the projection methods so that they use symmetric
quadrature rule and functions evaluated at the reference (non-oriented) points.