Intrepid2: add H(vol) hierarchical pyramids (#12136)
## Motivation
This PR adds an arbitrary-order, hierarchical basis for H(vol) on the pyramid.  This follows #12079, which added hierarchical bases for H(grad) on the pyramid.

Additionally, this PR adds "sub-basis inclusion" tests for both H(vol) and H(grad) on the pyramid, and orientation tests against H(grad) on the pyramid.

## Testing
This PR includes "sub-basis inclusion" tests for both H(vol) and H(grad) on the pyramid, and orientation tests against H(grad) on the pyramid.  It also adds a simple basis cardinality test for the new H(vol) basis.

Additionally, I have done offline comparison with the ESEAS implementation of this basis, testing up to 10th order, with excellent agreement.  I hope to include these tests with Intrepid2 soon; this has been prevented previously by the license for ESEAS, but they are changing the license to one that will allow inclusion.