Intrepid2: Removing Experimental namespace used for projections (part1) (#12141)
* Intepid2: Minor modifications to CellTools

We now allow to map ref cells in a physical space that is higher dimensional
(e.g. quadrilateral in a 3d physical space)
We added a batched version of matToReferenceSubcell, where the subcell ordinal is a vector.
This is useful when working on a batch of sides

* Intrepid2: initial work to remove Experimental namespace used for projections

We introduce the cmake option Intrepid2_KEEP_EXPERIMENTAL_NAMESPACE. When set,
we use the legacy namespace, otherwise the Experimental namespace is removed. The option is ON by default.
Some of the legacy functions are not compiled when the flag is turned off

* Panzer: work to remove the Intrepid2::Experimental namespace used for projections

Also some improvements to test_fe_assembly and test_fe_projection code