Belos: Fixed the CG multi-precision test for Tpetra
The block CG multi-precision test was not built or tested, plus it
had old Tpetra syntax and needed updating.  It was similar to the
block GMRES multi-precision test, so it is being replaced with that
test modified to call block CG.  Now this new multi-precision test
is being enabled if the float precision is enabled in Tpetra.

After reviewing the tests and examples, it was noticed that the old
Dummy files are still in the repository.  They are unnecessary and
are being removed.