Tpetra: device version of sort_and_merge_crs
This should increase the speed of MueLu by a nice amount
on GPU platforms : )
Making input parameters non const and calling the new
Kokkos Kernels function.
With the new interface to the sorting facilities in KK we should
be able to get faster TAFC.
Basically one should have expected that views coming out of the Kokkos
Kernels routine are of different sizes so their host counterparts need
to be resized ahead of deep_copy...
Adding more logic to handle issues related to Kokkos::resize
and ArrayRCP::resize.