Tpetra: device version of sort_crs_matrix
In transfer and fill complete, the sort of crs matrix data is
currently done on host using ArrayRCP, we are moving this call
to device and storing data in View. Data is subsequently copied
and moved back to host and ArrayRCP while other routines are being
moved on device as well until we can skip the use of ArrayRCP all
together.