teuchos: Add missing export annotation for windows
The `error_out` function can be referenced from other shared objects
(likely via instantiations of the class defined in the same header).
As such, it must be exported to avoid link errors on Windows. For
example, I see the following in a windows build of tpetra:
```
/opt/x86_64-w64-mingw32/bin/../lib/gcc/x86_64-w64-mingw32/9.1.0/../../../../x86_64-w64-mingw32/bin/ld: CMakeFiles/tpetra.dir/objects.a(Tpetra_Details_DeepCopyTeuchosTimerInjection.cpp.obj):Tpetra_Details_DeepCopyTeuchosTimerInjection.cpp:(.text+0x279): undefined reference to `Teuchos::error_out(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, bool)'
```
This PR fixes this by adding the appropriate export annotation.