Intrepid2: add hierarchical bases for H(div) pyramids (#12286)
@trilinos/intrepid2 

## Motivation
This PR adds an arbitrary-order, hierarchical basis for H(div) on the pyramid.  This follows #12079, which added hierarchical bases for H(grad) on the pyramid, and #12136, which added bases for H(vol).

This PR also adds support in CellGeometry for a pyramidal subdivision of regular hexahedral meshes, and support for a rudimentary node numbering within CellGeometry.  This is in support of new projection convergence tests, so far implemented for H(grad) and H(vol).  H(div) support for such projections actually requires H(curl) bases to be implemented as well; those will be coming in a subsequent PR.

Additionally, this PR adds orientation tests against H(div) on the pyramid.

## Testing
This PR has orientation and basis cardinality tests for the new H(div) basis.  The new CellGeometry features have some limited tests against them as well.  It also adds projection convergence tests for H(grad) and H(vol) on the pyramid.

Additionally, I have done offline comparison with the ESEAS implementation of this basis, testing up to 9th order, with excellent agreement.  (There were actually a couple of bugs I found in ESEAS's implementation, which I plan to submit fixes for there shortly.)  I hope to include these tests with Intrepid2 soon; this has been prevented previously by the license for ESEAS, but they are changing the license to one that will allow inclusion.