Properly declare imported 'm' library
As per modern CMake recommendations, want to inform CMake that 'm'
('libm') is an imported library outside of the project.  Necessary when
using LINK_LIBRARIES_ONLY_TARGETS to allow CMake to error-check target
dependencies (since this is now a target dependency and not a direct
library dependency).

User Support Ticket(s) or Story Referenced: TRILFRAME-601