teuchos: Add more missing Windows export annotations
This follows up PR #12307 to address another round of missing exports.
I didn't make any efforts to systematically try to indentify what should
be exported. Rather, these are the symbols that the compiler complains
about in a Trilinos build with sufficient enabled components to build
charon (which uses a fair number of Trilinos packages).